Fixed mantis bug #11044
git-svn-id: http://svn.ilias.de/svn/ilias/branches/Release_4_3_x_branch@42909 21b2c9ec-7c21-0410-8b45-9bfb8ed2bfc5

+++ Modules/TestQuestionPool/classes/class.assTextQuestion.php
@@ -137,7 +137,7 @@ class assTextQuestion extends assQuestion
 		{
 			$nextID = $ilDB->nextId('qpl_a_essay');
 			$ilDB->manipulateF("INSERT INTO qpl_a_essay (answer_id, question_fi, answertext, points) VALUES (%s, %s, %s, %s)",
-							   array("integer", "integer", "text", 'integer'),
+							   array("integer", "integer", "text", 'float'),
 							   array(
 								   $nextID,
 								   $this->getId(),
